requirements.txt
speechrecognition
gTTS
pygame
python-dotenv
google-generativeai
sqlalchemy
pymysql
streamlit
pydub